services/atomic/notification/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
fastapi==0.115.11
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.3
geographiclib==2.0
geopy==2.4.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mlxtend==0.23.4
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pika==1.3.2
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.4
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
# Added rows
fastapi
uvicorn
twilio
pika
python-dotenv
requests
